backend/app/requirements.txt
fastapi
uvicorn[standard]
pydantic
pycoingecko
pandas
numpy
scikit-learn
lightgbm
passlib[bcrypt]
pyjwt
yfinance